+++ extensions/todo/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @zhushanwen/pi-todo
 
+## 0.1.6
+
+### Patch Changes
+
+- 15b68f6: Fix evolve analyzer to find session files in project subdirectories, unify pi.extensions to ./index.ts
+
 ## 0.1.5
 
 ### Patch Changes